ui(refactor): split 'superdesk-highlights' module.
