+++ scripts/superdesk-highlights/controllers/index.js
@@ -1,0 +1,2 @@
+export { HighlightsSettings } from './HighlightsSettings';
+export { HighlightsConfig } from './HighlightsConfig';
