+++ scripts/superdesk-highlights/directives/HighlightsConfigModal.js
@@ -1,0 +1,8 @@
+export function HighlightsConfigModal() {
+    return {
+        require: '^sdHighlightsConfig',
+        templateUrl: 'scripts/superdesk-highlights/views/highlights_config_modal.html',
+        link: function(scope, elem, attrs, ctrl) {
+        }
+    };
+}
