+++ scripts/superdesk-highlights/directives/HighlightsInfo.js
@@ -1,0 +1,9 @@
+HighlightsInfo.$inject = [];
+export function HighlightsInfo() {
+    return {
+        scope: {
+            item: '=item'
+        },
+        templateUrl: 'scripts/superdesk-highlights/views/highlights_info_directive.html'
+    };
+}
