+++ scripts/superdesk-highlights/directives/index.js
@@ -1,0 +1,10 @@
+export { MarkHighlightsDropdown } from './MarkHighlightsDropdown';
+export { MultiMarkHighlightsDropdown } from './MultiMarkHighlightsDropdown';
+export { HighlightsInfo } from './HighlightsInfo';
+export { HighlightsTitle } from './HighlightsTitle';
+export { SearchHighlights } from './SearchHighlights';
+export { PackageHighlightsDropdown } from './PackageHighlightsDropdown';
+export { HighlightsLabel } from './HighlightsLabel';
+export { CreateHighlightsButton } from './CreateHighlightsButton';
+export { HighlightsConfigModal } from './HighlightsConfigModal';
+export { HighlightsReactDropdown } from './HighlightsReactDropdown';
