+++ scripts/superdesk-highlights/services/index.js
@@ -1,0 +1,1 @@
+export { HighlightsService } from './HighlightsService';
