+++ scripts/superdesk-highlights/tests/highlights.spec.js
